+++ src/app/features/events/events.component.ts
@@ -40,7 +40,7 @@ export class EventsComponent implements OnInit, OnDestroy {
   displayedColumns: string[] = [
       'status', 'id', 'timestamp',
       'image', 'collapsed-info', 'name', 'categories', 'dateIni',
-      'actions3', 'social'
+      'actions4', 'social'
 ];
   private listOfObservers: Array<Subscription> = [];
   private currentUser: IUser;
@@ -128,6 +128,11 @@ export class EventsComponent implements OnInit, OnDestroy {
         });
       }
     });
+  }
+
+  deleteForeverElement(event: IEvent): void {
+    this.logSrv.info(`deleteForeverElement: ${JSON.stringify(event.id)}`);
+    this.eventSrv.deleteForeverEvent(event);
   }
 
   public addItem(): void {